Move AltResult to primitives.
