+++ Cargo.toml
@@ -48,6 +48,7 @@ futures-util = "0.3.30"
 hex = "0.4.3"
 libp2p = "0.53.2"
 libp2p-stream = "0.1.0-alpha.1"
+libp2p-identity = "0.2.9"
 multiaddr = "0.18.1"
 multibase = "0.9.1"
 near-jsonrpc-client = "0.8.0"
