+++ crates/sdk/libs/near/Cargo.toml
@@ -10,4 +10,5 @@ license.workspace = true
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
 calimero-sdk = { path = "../../" }
+calimero-primitives = { path = "../../../primitives" }
