+++ crates/sdk/libs/near/src/jsonrpc.rs
@@ -1,24 +1,6 @@
 use calimero_sdk::env;
 use serde::de::DeserializeOwned;
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize)]
-#[serde(remote = "Result")]
-pub enum ResultAlt<T, E> {
-    #[serde(rename = "result")]
-    Ok(T),
-    #[serde(rename = "error")]
-    Err(E),
-}
-
-impl<T, E> From<ResultAlt<T, E>> for Result<T, E> {
-    fn from(result: ResultAlt<T, E>) -> Self {
-        match result {
-            ResultAlt::Ok(value) => Ok(value),
-            ResultAlt::Err(err) => Err(err),
-        }
-    }
-}
+use serde::Deserialize;
 
 pub(crate) struct Client {
     url: String,
@@ -55,7 +37,7 @@ pub struct Response<T: DeserializeOwned, E: DeserializeOwned> {
     pub jsonrpc: Option<String>,
     pub id: String,
 
-    #[serde(with = "ResultAlt", flatten)]
+    #[serde(with = "calimero_primitives::common::ResultAlt", flatten)]
     pub data: Result<T, RpcError<E>>,
 }
